Generate objects/dependencies using absolute path
When adding source files, relative path like ".." might be used. If
pattern rule searches for file using this relative path, the target
object/dependency file may end up outside the build directory. To avoid
such kind of troubles, objects/dependencies in the build directory
are all generated using the absolute path of the source files.

With this approach when the whole project is moved to a different path,
make will rebuild all the sources because they now have absolute path.